Show only rgb-template in previous mode
